DL/requirements_local_conda.txt
absl-py
astunparse
cachetools
certifi
chardet
click
cycler
Flask
gast
google-api-core
google-auth
google-auth-oauthlib
google-cloud-core
google-cloud-storage
google-pasta
google-resumable-media
googleapis-common-protos
grpcio
h5py
idna
itsdangerous
Jinja2
joblib
Keras-Preprocessing
kiwisolver
Markdown
MarkupSafe
matplotlib
numpy
oauthlib
Pillow
protobuf
pyasn1
pyasn1-modules
pyparsing
python-dateutil
pytz
requests
requests-oauthlib
rsa
scikit-learn
scipy
six
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor
threadpoolctl
urllib3
Werkzeug
wrapt
